requirements.txt
flask
gunicorn
openai
python-dotenv
twilio